FEATURE: Add user_promoted trigger and event hook
Added the new user_promoted trigger which has the following fields:

* restricted_group:       Will only run the script if the user being promoted
                          is in the specified group.
* trust_level_transition: Will only run the script if old to new trust
                          level matches what is specified here (e.g. 1
                          to 2), or for all trust level promotions.

We use the :user_promoted event from core for this new trigger. We do
not fire the automation trigger if the "promotion" is actually a
demotion e.g. 4 -> 1. We may want a separate event in core for this at
some point.