DEV: normalise using custom fields with automations
Setting custom fields on topic/post/user is a very classic use case. It opens various efficient ways to react on event and look if an automation is associated with it.

Note that this commit also renames the existing discourse_automation_id on topic_custom_fields to discourse_automation_ids to make explicit it's now an array.