FEATURE: adds auto_responder script
The auto responder script allows to define a list of keys/values which will be used to react to a post content. If a post contains KEY, then a reply using VALUE will be made to the post.

Note that this commit also introduces the ability for the post_created_edited trigger to limit its action to specific trust_levels creators.