Add add_user_to_group_through_custom_field script
It accepts a Recurring trigger, and takes the name of a User Custom Field containing a group name.
On each run, it ensures that each user belongs to the group name given by their UCF.

In other words, it designates a certain User Custom Field to act as a "pointer" to a group that the user should belong to, and adds users to that group as needed.