FIX: prevents saving an enabled automation with no trigger (#153)
Under specific circumstances an automation could be saved as enabled without a trigger, which could lead to the automation staying enabled after adding an uncompleted trigger.

This could eventually me moved into some automation validation, but that can be quite tricky to do given the chicken-egg situation of script needing triggers, triggers needs automation, ... and other added complexity like validity of relationships (fields of the scripts/triggers, enabled or not...).

The enabled checkbox has been moved to the bottom and more importantly is only showing if a trigger is set. We could need a different pattern for this as in the past having this field at the bottom led to people not seeing it, but this is too much complexity for now.