FEATURE: Add script to modify user group memberships through badges
This change adds a new script, `user_group_membership_through_badge`
which provides the option to automate user group memberships through
badges.

Given a badge name and a group, this script adds users with the
specified badged to the the given group on login or on a recurring
schedule.

It also has an optional flag which when enabled ensures only users with the
specified badge are members of the given group