FEATURE: Additional options for pm_created and post_created_edited triggers
This commit introduces some additional options to make the
pm_created and post_created_edited triggers more useable as
auto-responders inside group inboxes.

* restricted_group - Will only call the trigger if the topic is a PM
sent to the specified group.
* ignore_automated - If the post/topic is being created from an incoming
email, and it is from a noreply address or otherwise automated, then it
will not be considered valid for the autoresponder to reply to.
* ignore_group_members - When combined with restricted_group, the
autoresponder will not fire the trigger if the post/topic is being
created by a member of the group.

This commit also does some minor refactors and makes some UI
and copy improvements.